refactored demos/appyx-navigation/common into a multiplatform library used by demos/appyx-navigation/android
Disabled DatingCardsNode in ContainerNode as image loading solution needs to be provided in a multiplatform-supported way
