+++ demos/appyx-navigation/common/build.gradle.kts
@@ -1,0 +1,79 @@
+plugins {
+    kotlin("multiplatform")
+    id("org.jetbrains.compose")
+    id("com.android.library")
+    id("appyx-publish-multiplatform")
+}
+
+kotlin {
+    android {
+        publishLibraryVariants("release")
+    }
+    jvm("desktop") {
+        compilations.all {
+            kotlinOptions.jvmTarget = libs.versions.jvmTarget.get()
+        }
+    }
+    js(IR) {
+        // Adding moduleName as a workaround for this issue: https://youtrack.jetbrains.com/issue/KT-51942
+        moduleName = "demo-appyx-navigation-common"
+        browser()
+    }
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                api(compose.runtime)
+                api(compose.foundation)
+                api(compose.material3)
+                api(project(":utils:multiplatform"))
+                implementation(libs.kotlinx.serialization.json)
+                api(project(":utils:customisations"))
+                api(project(":appyx-interactions:appyx-interactions"))
+                api(project(":appyx-navigation:common"))
+                implementation(project(":appyx-components:stable:backstack:backstack"))
+                implementation(project(":appyx-components:experimental:cards:cards"))
+                implementation(project(":appyx-components:experimental:promoter:promoter"))
+                implementation(project(":appyx-components:stable:spotlight:spotlight"))
+            }
+        }
+        val commonTest by getting {
+            dependencies {
+                implementation(kotlin("test"))
+            }
+        }
+        val androidMain by getting {
+            dependencies {
+                api(libs.androidx.appcompat)
+                api(libs.androidx.core)
+                implementation(libs.coil.compose)
+                implementation(libs.androidx.activity.compose)
+            }
+        }
+        val desktopMain by getting {
+            dependencies {
+                api(compose.preview)
+            }
+        }
+        val jsMain by getting {
+            dependencies {
+                implementation(npm("uuid", libs.versions.uuid.get()))
+            }
+        }
+    }
+}
+
+android {
+    namespace = "com.bumble.appyx.demos.appyxnavigation"
+    compileSdk = libs.versions.androidCompileSdk.get().toInt()
+    sourceSets["main"].manifest.srcFile("src/androidMain/AndroidManifest.xml")
+    defaultConfig {
+        minSdk = libs.versions.androidMinSdk.get().toInt()
+        targetSdk = libs.versions.androidTargetSdk.get().toInt()
+    }
+    buildFeatures {
+        compose = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = libs.versions.composeCompiler.get()
+    }
+}
