+++ demos/appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/node/backstack/debug/BackstackDebugNode.kt
@@ -1,15 +1,13 @@
-package com.bumble.appyx.navigation.node.backstack
+package com.bumble.appyx.navigation.node.backstack.debug
 
-import android.os.Parcelable
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
@@ -21,50 +19,48 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.bumble.appyx.components.backstack.BackStack
 import com.bumble.appyx.components.backstack.BackStackModel
-import com.bumble.appyx.components.backstack.BackStackModel.State
 import com.bumble.appyx.components.backstack.operation.newRoot
 import com.bumble.appyx.components.backstack.operation.pop
 import com.bumble.appyx.components.backstack.operation.push
 import com.bumble.appyx.components.backstack.operation.replace
-import com.bumble.appyx.interactions.core.ui.MotionController
-import com.bumble.appyx.interactions.core.ui.context.TransitionBounds
-import com.bumble.appyx.interactions.core.ui.context.UiContext
-import com.bumble.appyx.interactions.core.ui.gesture.GestureFactory
-import com.bumble.appyx.interactions.core.ui.gesture.GestureSettleConfig
+import com.bumble.appyx.components.backstack.ui.slider.BackStackSlider
 import com.bumble.appyx.navigation.colors
 import com.bumble.appyx.navigation.composable.Children
+import com.bumble.appyx.navigation.composable.KnobControl
 import com.bumble.appyx.navigation.modality.BuildContext
 import com.bumble.appyx.navigation.node.Node
 import com.bumble.appyx.navigation.node.ParentNode
+import com.bumble.appyx.navigation.node.backstack.debug.BackstackDebugNode.InteractionTarget
 import com.bumble.appyx.navigation.node.node
-import com.bumble.appyx.navigation.ui.TextButton
 import com.bumble.appyx.navigation.ui.appyx_dark
-import gestureModifier
-import kotlinx.parcelize.Parcelize
-import kotlin.random.Random
+import com.bumble.appyx.utils.multiplatform.Parcelable
+import com.bumble.appyx.utils.multiplatform.Parcelize
 
-
-class BackStackNode(
+class BackstackDebugNode(
     buildContext: BuildContext,
-    motionController: (UiContext) -> MotionController<InteractionTarget, State<InteractionTarget>>,
-    gestureFactory: (TransitionBounds) -> GestureFactory<InteractionTarget, State<InteractionTarget>> = {
-        GestureFactory.Noop()
-    },
-    gestureSettleConfig: GestureSettleConfig = GestureSettleConfig(),
-    private val isMaxSize: Boolean = false,
     private val backStack: BackStack<InteractionTarget> = BackStack(
         model = BackStackModel(
             initialTargets = listOf(InteractionTarget.Child(1)),
-            savedStateMap = buildContext.savedStateMap
+            savedStateMap = buildContext.savedStateMap,
         ),
-        motionController = motionController,
-        gestureFactory = gestureFactory,
-        gestureSettleConfig = gestureSettleConfig,
+        motionController = { BackStackSlider(it) }
     )
-) : ParentNode<BackStackNode.InteractionTarget>(
+) : ParentNode<InteractionTarget>(
     buildContext = buildContext,
-    interactionModel = backStack,
+    interactionModel = backStack
 ) {
+
+    init {
+        backStack.push(InteractionTarget.Child(2))
+        backStack.push(InteractionTarget.Child(3))
+        backStack.push(InteractionTarget.Child(4))
+        backStack.push(InteractionTarget.Child(5))
+        backStack.replace(InteractionTarget.Child(6))
+        backStack.pop()
+        backStack.pop()
+        backStack.newRoot(InteractionTarget.Child(1))
+    }
+
     sealed class InteractionTarget : Parcelable {
         @Parcelize
         class Child(val index: Int) : InteractionTarget()
@@ -78,16 +74,9 @@ class BackStackNode(
                 Box(
                     modifier = Modifier
                         .fillMaxSize()
-                        .then(
-                            if (isMaxSize) {
-                                Modifier
-                            } else {
-                                Modifier.clip(RoundedCornerShape(5))
-                            }
-                        )
+                        .clip(RoundedCornerShape(5))
                         .background(backgroundColor)
                         .padding(24.dp)
-                        .gestureModifier(backStack, interactionTarget.index.toString())
                 ) {
                     Text(
                         text = interactionTarget.index.toString(),
@@ -99,6 +88,7 @@ class BackStackNode(
             }
         }
 
+    @ExperimentalMaterialApi
     @Composable
     override fun View(modifier: Modifier) {
         Column(
@@ -106,41 +96,16 @@ class BackStackNode(
                 .fillMaxWidth()
                 .background(appyx_dark)
         ) {
+            KnobControl(onValueChange = {
+                backStack.setNormalisedProgress(it)
+            })
             Children(
-                clipToBounds = true,
                 interactionModel = backStack,
                 modifier = Modifier
-                    .weight(0.9f)
                     .fillMaxSize()
                     .background(appyx_dark)
-                    .then(
-                        if (isMaxSize) {
-                            Modifier.padding(bottom = 16.dp)
-                        } else {
-                            Modifier.padding(16.dp)
-                        }
-                    ),
+                    .padding(16.dp),
             )
-            Row(
-                modifier = Modifier
-                    .weight(0.1f)
-                    .fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceEvenly
-            ) {
-                TextButton(text = "Push") {
-                    backStack.push(InteractionTarget.Child(Random.nextInt(20)))
-                }
-                TextButton(text = "Pop") {
-                    backStack.pop()
-                }
-                TextButton(text = "Replace") {
-                    backStack.replace(InteractionTarget.Child(Random.nextInt(20)))
-                }
-                TextButton(text = "New root") {
-                    backStack.newRoot(InteractionTarget.Child(Random.nextInt(20)))
-                }
-            }
         }
     }
 }
-
