+++ demos/appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/node/spotlight/debug/SpotlightDebugNode.kt
@@ -1,18 +1,13 @@
-package com.bumble.appyx.navigation.node.spotlight
+package com.bumble.appyx.navigation.node.spotlight.debug
 
-import android.os.Parcelable
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.spring
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Button
+import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
@@ -28,33 +23,43 @@ import com.bumble.appyx.components.spotlight.operation.first
 import com.bumble.appyx.components.spotlight.operation.last
 import com.bumble.appyx.components.spotlight.operation.next
 import com.bumble.appyx.components.spotlight.operation.previous
-import com.bumble.appyx.components.spotlight.operation.updateElements
 import com.bumble.appyx.components.spotlight.ui.slider.SpotlightSlider
 import com.bumble.appyx.navigation.colors
 import com.bumble.appyx.navigation.composable.Children
+import com.bumble.appyx.navigation.composable.KnobControl
 import com.bumble.appyx.navigation.modality.BuildContext
 import com.bumble.appyx.navigation.node.Node
 import com.bumble.appyx.navigation.node.ParentNode
 import com.bumble.appyx.navigation.node.node
-import com.bumble.appyx.navigation.node.spotlight.SpotlightNode.InteractionTarget
+import com.bumble.appyx.navigation.node.spotlight.debug.SpotlightDebugNode.InteractionTarget
 import com.bumble.appyx.navigation.ui.appyx_dark
-import kotlinx.parcelize.Parcelize
+import com.bumble.appyx.utils.multiplatform.Parcelable
+import com.bumble.appyx.utils.multiplatform.Parcelize
 
-class SpotlightNode(
+class SpotlightDebugNode(
     buildContext: BuildContext,
     private val spotlight: Spotlight<InteractionTarget> = Spotlight(
         model = SpotlightModel(
-            items = List(7) { InteractionTarget.Child(it) },
+            items = List(7) { InteractionTarget.Child(it + 1) },
             initialActiveIndex = 0f,
             savedStateMap = buildContext.savedStateMap
         ),
-        motionController = { SpotlightSlider(it) }
+        motionController = { SpotlightSlider(it) },
+        isDebug = true
     )
 ) : ParentNode<InteractionTarget>(
     buildContext = buildContext,
     interactionModel = spotlight
 ) {
-    private val newItems = List(7) { InteractionTarget.Child(it * 3) }
+
+    init {
+        spotlight.next()
+        spotlight.next()
+        spotlight.next()
+        spotlight.previous()
+        spotlight.last()
+        spotlight.first()
+    }
 
     sealed class InteractionTarget : Parcelable {
         @Parcelize
@@ -82,50 +87,24 @@ class SpotlightNode(
             }
         }
 
+    @ExperimentalMaterialApi
     @Composable
     override fun View(modifier: Modifier) {
         Column(
-            modifier = modifier
-                .fillMaxSize()
+            modifier
+                .fillMaxWidth()
                 .background(appyx_dark)
         ) {
+            KnobControl(onValueChange = {
+                spotlight.setNormalisedProgress(it)
+            })
             Children(
                 interactionModel = spotlight,
-                modifier = Modifier
-                    .padding(
-                        horizontal = 64.dp,
-                        vertical = 12.dp
-                    )
-                    .weight(0.9f)
+                modifier = Modifier.padding(
+                    horizontal = 64.dp,
+                    vertical = 12.dp
+                )
             )
-
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .weight(0.1f),
-                horizontalArrangement = Arrangement.SpaceEvenly
-            ) {
-                Button(onClick = {
-                    spotlight.updateElements(
-                        newItems.shuffled(),
-                        animationSpec = spring(stiffness = Spring.StiffnessVeryLow / 20)
-                    )
-                }) {
-                    Text("New")
-                }
-                Button(onClick = { spotlight.first() }) {
-                    Text("First")
-                }
-                Button(onClick = { spotlight.previous(spring(stiffness = Spring.StiffnessLow)) }) {
-                    Text("Prev")
-                }
-                Button(onClick = { spotlight.next(spring(stiffness = Spring.StiffnessMedium)) }) {
-                    Text("Next")
-                }
-                Button(onClick = { spotlight.last() }) {
-                    Text("Last")
-                }
-            }
         }
     }
 }
