+++ demos/appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/ui/Type.kt
