+++ settings.gradle.kts
@@ -47,7 +47,9 @@ include(
     ":demos:appyx-interactions:android",
     ":demos:appyx-interactions:desktop",
     ":demos:appyx-interactions:web",
+    ":demos:appyx-navigation:common",
     ":demos:appyx-navigation:android",
+    ":demos:appyx-navigation:desktop",
     ":demos:common",
     ":demos:navigation-compose",
     ":demos:mkdocs:appyx-interactions:interactions:sample1:web",
